move sourmash._minhash to sourmash.minhash
