+++ doc/conf.py
@@ -297,4 +297,4 @@ texinfo_documents = [
 # If true, do not generate a @detailmenu in the "Top" node's menu.
 #texinfo_no_detailmenu = False
 
-autodoc_mock_imports = ["sourmash._minhash"]
+autodoc_mock_imports = ["sourmash.minhash"]
