+++ sourmash/cli/compute.py
@@ -28,7 +28,7 @@ Please see -h for all of the options as well as more detailed help.
 
 from argparse import FileType
 
-from sourmash._minhash import get_minhash_default_seed
+from sourmash.minhash import get_minhash_default_seed
 from sourmash.cli.utils import add_construct_moltype_args
