+++ sourmash/nodegraph.py
@@ -7,7 +7,7 @@ from tempfile import NamedTemporaryFile
 
 from ._compat import string_types, range_type
 from ._lowlevel import ffi, lib
-from ._minhash import to_bytes, MinHash
+from .minhash import to_bytes, MinHash
 from .utils import RustObject, rustcall, decode_str
 from .exceptions import SourmashError
