+++ sourmash/search.py
@@ -4,7 +4,7 @@ import sys
 
 from .logging import notify, error
 from .signature import SourmashSignature
-from ._minhash import get_max_hash_for_scaled
+from .minhash import get_max_hash_for_scaled
 
 
 # generic SearchResult.
