+++ sourmash/sig/__main__.py
@@ -12,9 +12,9 @@ import sourmash
 import copy
 from sourmash.sourmash_args import FileOutput
 
-from ..logging import set_quiet, error, notify, set_quiet, print_results, debug
-from .. import sourmash_args
-from .._minhash import get_max_hash_for_scaled
+from sourmash.logging import set_quiet, error, notify, set_quiet, print_results, debug
+from sourmash import sourmash_args
+from sourmash.minhash import get_max_hash_for_scaled
 
 usage='''
 sourmash signature <command> [<args>] - manipulate/work with signature files.
